Phase 3.1: port copyInto + Support.java from reverse (array transfer)
Consolidates the first slice of plan/ArrayTransferPhase3.md onto this
branch, ahead of the push/pull benchmarking and parallel-support work.

- Extract org.jpype.internal.Support (collect/collectRectangular/unpack/
  assemble) out of JPypeContext, matching reverse's split. C++ now
  resolves it once and calls static methods instead of round-tripping
  through the context instance; also fixes a latent bug where
  assemble()'s null-check tested the wrong method-ID member.
- Port JArray.copyInto() (bulk Java-array -> caller-buffer pull): a
  contiguous fast path via a single Get<Type>ArrayRegion call, and a
  GetPrimitiveArrayCritical + stride-walk general path for stepped
  sources/non-contiguous destinations. copyElements/getItemSize already
  existed per-type on this branch, just unused for this; only the
  strided-fallback helper and JPArray::copyInto itself were new.
- Fix JPIntType::getItemSize() returning sizeof(jfloat) instead of
  sizeof(jint), found while touching this code (harmless in practice,
  both are 4 bytes, but wrong).
- Add test_arrayCopyInto.py (fast path, non-contiguous dest, stepped
  source, size/itemsize mismatches, non-primitive array, empty array,
  read-only dest).
- Add project/benchmark/jpype/arraytransfer.py: the 4 comparison models
  from reverse's benchmark/arraybench/ (copyInto bulk, direct-buffer-
  shared, slicing, 2D bulk transfer), rewritten Python-driven instead of
  ported verbatim -- the original harness drives Java-to-Python through
  reverse's bridge (MainInterpreter/Script/python.lang), which doesn't
  exist on this branch.

+++ native/common/jp_inttype.cpp
@@ -345,7 +345,7 @@ const char* JPIntType::getBufferFormat()
 
 Py_ssize_t JPIntType::getItemSize()
 {
-	return sizeof (jfloat);
+	return sizeof (jint);
 }
 
 void JPIntType::copyElements(JPJavaFrame &frame, jarray a, jsize start, jsize len,
